refactor: split telegram media and app wiring
